# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: GameHub API
  description: GameHub游戏社区平台API文档
  version: 1.0.0
  contact:
    name: GameHub开发团队
    email: dev@gamehub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发服务器
  - url: https://api.gamehub.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /reviews/{id}/like:
    post:
      tags:
        - 评测
      summary: 点赞评测
      description: 点赞游戏评测
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 点赞成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      reviewId:
                        type: string
                      likes:
                        type: integer
                      liked:
                        type: boolean
...
  /community/posts/{id}/like:
    post:
      tags:
        - 社区
      summary: 点赞帖子
      description: 点赞社区帖子
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 点赞成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      postId:
                        type: string
                      likes:
                        type: integer
                      liked:
                        type: boolean
...
  /community/reviews/{id}/like:
    post:
      tags:
        - 社区
      summary: 点赞社区评测
      description: 点赞社区评测
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 点赞成功
...
  /like:
    post:
      tags:
        - 点赞
      summary: 添加点赞
      description: 对目标内容添加点赞
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetType:
                  type: string
                  enum: [review, news, community_post, comment]
                targetId:
                  type: string
      responses:
        '201':
          description: 点赞成功
    delete:
      tags:
        - 点赞
      summary: 取消点赞
      description: 取消对目标内容的点赞
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetType:
                  type: string
                targetId:
                  type: string
      responses:
        '200':
          description: 取消点赞成功
  /like/status:
    get:
      tags:
        - 点赞
      summary: 点赞状态
      description: 检查是否已点赞
      security:
        - bearerAuth: []
      parameters:
        - name: targetType
          in: query
          required: true
          schema:
            type: string
        - name: targetId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
...
  /comments/{id}/like:
    post:
      tags:
        - 评论
      summary: 点赞评论
      description: 点赞评论
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 点赞成功
...
  /like/target:
    get:
      tags:
        - 点赞
      summary: 获取目标点赞列表
      description: 获取指定目标的所有点赞用户
      parameters:
        - name: targetType
          in: query
          required: true
          schema:
            type: string
        - name: targetId
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 获取成功
  /like/user:
    get:
      tags:
        - 点赞
      summary: 获取用户点赞列表
      description: 获取指定用户的点赞历史
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 获取成功
  /like/stats:
    get:
      tags:
        - 点赞
      summary: 点赞统计
      description: 获取点赞统计数据
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
...
  /gamification/stats:
    get:
      tags:
        - 游戏化
      summary: 获取游戏化统计
      description: 获取当前用户的经验值和积分统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 获取成功
  /gamification/xp/history:
    get:
      tags:
        - 游戏化
      summary: 经验值历史
      description: 获取经验值变动历史
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
  /gamification/leaderboard/{type}:
    get:
      tags:
        - 游戏化
      summary: 游戏化排行榜
      description: 获取经验值/积分排行榜
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [xp, points, level]
      responses:
        '200':
          description: 获取成功
...
  /guides/{id}/like:
    post:
      tags:
        - 攻略
      summary: 点赞攻略
      description: 点赞攻略指南
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 点赞成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...